add compile helper
