+++ src/main/java/edu/byu/cs/autograder/Grader.java
@@ -1,6 +1,7 @@
 package edu.byu.cs.autograder;
 
 import edu.byu.cs.analytics.CommitAnalytics;
+import edu.byu.cs.autograder.compile.CompileHelper;
 import edu.byu.cs.autograder.git.GitHelper;
 import edu.byu.cs.autograder.score.Scorer;
 import edu.byu.cs.model.*;
@@ -31,6 +32,8 @@ public abstract class Grader implements Runnable {
 
     private final GitHelper gitHelper;
 
+    private final CompileHelper compileHelper;
+
     protected final GradingContext gradingContext;
 
     protected Observer observer;
@@ -54,6 +57,7 @@ public abstract class Grader implements Runnable {
                 new GradingContext(netId, phase, phasesPath, stagePath, repoUrl, stageRepo, requiredCommits, observer);
         this.dbHelper = new DatabaseHelper(salt, gradingContext);
         this.gitHelper = new GitHelper(gradingContext);
+        this.compileHelper = new CompileHelper(gradingContext);
     }
 
     public void run() {
@@ -62,13 +66,8 @@ public abstract class Grader implements Runnable {
             // FIXME: remove this sleep. currently the grader is too quick for the client to keep up
             Thread.sleep(1000);
             int numCommits = gitHelper.setUp();
-            verifyProjectStructure();
-
             dbHelper.setUp();
-
-            modifyPoms();
-
-            packageRepo();
+            compileHelper.compile();
 
             RubricConfig rubricConfig = DaoService.getRubricConfigDao().getRubricConfig(gradingContext.phase());
             Rubric.Results qualityResults = null;
@@ -132,22 +131,6 @@ public abstract class Grader implements Runnable {
         };
     }
 
-    private void modifyPoms() {
-        File oldRootPom = new File(gradingContext.stageRepo(), "pom.xml");
-        File oldServerPom = new File(gradingContext.stageRepo(), "server/pom.xml");
-        File oldClientPom = new File(gradingContext.stageRepo(), "client/pom.xml");
-        File oldSharedPom = new File(gradingContext.stageRepo(), "shared/pom.xml");
-
-        File newRootPom = new File(gradingContext.phasesPath(), "pom/pom.xml");
-        File newServerPom = new File(gradingContext.stageRepo(), "pom/server/pom.xml");
-        File newClientPom = new File(gradingContext.stageRepo(), "pom/client/pom.xml");
-        File newSharedPom = new File(gradingContext.stageRepo(), "pom/shared/pom.xml");
-
-        FileUtils.copyFile(oldRootPom, newRootPom);
-        FileUtils.copyFile(oldServerPom, newServerPom);
-        FileUtils.copyFile(oldClientPom, newClientPom);
-        FileUtils.copyFile(oldSharedPom, newSharedPom);
-    }
 
     /**
      * Runs quality checks on the student's code
@@ -167,61 +150,6 @@ public abstract class Grader implements Runnable {
                 rubricConfig.quality().points(), null, quality.results());
     }
 
-    /**
-     * Verifies that the project is structured correctly. The project should be at the top level of the git repository,
-     * which is checked by looking for a pom.xml file
-     */
-    private void verifyProjectStructure() throws GradingException {
-        File pomFile = new File(gradingContext.stageRepo(), "pom.xml");
-        if (!pomFile.exists()) {
-            observer.notifyError("Project is not structured correctly. Your project should be at the top level of your git repository.");
-            throw new GradingException("No pom.xml file found");
-        }
-    }
-
-
-    /**
-     * Packages the student repo into a jar
-     */
-    protected void packageRepo() throws GradingException {
-        observer.update("Packaging repo...");
-
-        observer.update("  Running maven package command...");
-        ProcessBuilder processBuilder = new ProcessBuilder();
-        processBuilder.directory(gradingContext.stageRepo());
-        processBuilder.command("mvn", "package", "-DskipTests");
-        try {
-            ProcessUtils.ProcessOutput output = ProcessUtils.runProcess(processBuilder, 90000); //90 seconds
-            if (output.statusCode() != 0) {
-                throw new GradingException("Failed to package repo: ", getMavenError(output.stdOut()));
-            }
-        } catch (ProcessUtils.ProcessException ex) {
-            throw new GradingException("Failed to package repo", ex);
-        }
-
-        observer.update("Successfully packaged repo");
-    }
-
-    /**
-     * Retrieves maven error output from maven package stdout
-     *
-     * @param output A string containing maven standard output
-     * @return A string containing maven package error lines
-     */
-    private String getMavenError(String output) {
-        StringBuilder builder = new StringBuilder();
-        for (String line : output.split("\n")) {
-            if (line.contains("[ERROR] -> [Help 1]")) {
-                break;
-            }
-
-            if(line.contains("[ERROR]")) {
-                String trimLine = line.replace(gradingContext.stageRepo().getAbsolutePath(), "");
-                builder.append(trimLine).append("\n");
-            }
-        }
-        return builder.toString();
-    }
 
     /**
      * Run the unit tests written by the student. This approach is destructive as it will delete non-unit tests
